pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hl7_fhir_tool"
version = "0.1.0"
description = "HL7 v2 and FHIR viewer/transformer"
authors = [{ name = "Patrick Beach" }]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["hl7", "fhir", "interoperability", "healthcare"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "hl7apy>=1.3.4",
    "fhir.resources>=7.1.0",   # FHIR R4
    "lxml>=5.2.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "mypy>=1.11",
    "ruff>=0.6.0",
    "types-PyYAML",
]

# Expose a CLI without adding new runtime deps.
[project.scripts]
hl7-fhir = "hl7_fhir_tool.cli:main"

[project.urls]
Homepage = "https://github.com/shaolinpat/hl7_fhir_tool"
Issues = "https://github.com/shaolinpat/hl7_fhir_tool/issues"

[tool.setuptools.packages.find]
where = ["src"]

# Include any non-Python files specified in MANIFEST.in if you add one
[tool.setuptools]
include-package-data = true

[tool.ruff]
line-length = 100
target-version = "py311"

# Keep defaults minimal; you can expand later
src = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --cov=hl7_fhir_tool --cov-branch --cov-report=term-missing"